Changed type int to size_t

+++ main.cpp
@@ -10,9 +10,9 @@ bool palindrome(string sentence)
 {
     int x=0;
     sentence.erase(remove(sentence.begin(), sentence.end(), ' '), sentence.end());
-    int length=sentence.length();
+    size_t length=sentence.length();
         if(length==0){cout<<"No sentence was given.";exit(0);}
-            for(int i=0; i<(length/2) ;i++)
+            for(size_t i=0; i<(length/2) ;i++)
             {
                 if(sentence[x]==sentence[length-x-1])
                 x++;
